+++ tests/api_connexion/test_auth.py
@@ -16,15 +16,15 @@
 # under the License.
 from __future__ import annotations
 
-from base64 import b64encode
+from unittest.mock import patch
 
 import pytest
-from flask_login import current_user
 
+from airflow.auth.managers.simple.simple_auth_manager import SimpleAuthManager
+from airflow.auth.managers.simple.user import SimpleAuthManagerUser
 from tests.test_utils.api_connexion_utils import assert_401
 from tests.test_utils.config import conf_vars
 from tests.test_utils.db import clear_db_pools
-from tests.test_utils.www import client_with_login
 
 pytestmark = [pytest.mark.db_test, pytest.mark.skip_if_database_isolation_mode]
 
@@ -33,101 +33,6 @@ class BaseTestAuth:
     @pytest.fixture(autouse=True)
     def set_attrs(self, minimal_app_for_api):
         self.app = minimal_app_for_api
-
-        sm = self.app.appbuilder.sm
-        tester = sm.find_user(username="test")
-        if not tester:
-            role_admin = sm.find_role("Admin")
-            sm.add_user(
-                username="test",
-                first_name="test",
-                last_name="test",
-                email="test@fab.org",
-                role=role_admin,
-                password="test",
-            )
-
-
-class TestBasicAuth(BaseTestAuth):
-    @pytest.fixture(autouse=True, scope="class")
-    def with_basic_auth_backend(self, minimal_app_for_api):
-        from airflow.www.extensions.init_security import init_api_auth
-
-        old_auth = getattr(minimal_app_for_api, "api_auth")
-
-        try:
-            with conf_vars(
-                {("api", "auth_backends"): "airflow.providers.fab.auth_manager.api.auth.backend.basic_auth"}
-            ):
-                init_api_auth(minimal_app_for_api)
-                yield
-        finally:
-            setattr(minimal_app_for_api, "api_auth", old_auth)
-
-    def test_success(self):
-        token = "Basic " + b64encode(b"test:test").decode()
-        clear_db_pools()
-
-        with self.app.test_client() as test_client:
-            response = test_client.get("/api/v1/pools", headers={"Authorization": token})
-            assert current_user.email == "test@fab.org"
-
-        assert response.status_code == 200
-        assert response.json == {
-            "pools": [
-                {
-                    "name": "default_pool",
-                    "slots": 128,
-                    "occupied_slots": 0,
-                    "running_slots": 0,
-                    "queued_slots": 0,
-                    "scheduled_slots": 0,
-                    "deferred_slots": 0,
-                    "open_slots": 128,
-                    "description": "Default pool",
-                    "include_deferred": False,
-                },
-            ],
-            "total_entries": 1,
-        }
-
-    @pytest.mark.parametrize(
-        "token",
-        [
-            "basic",
-            "basic ",
-            "bearer",
-            "test:test",
-            b64encode(b"test:test").decode(),
-            "bearer ",
-            "basic: ",
-            "basic 123",
-        ],
-    )
-    def test_malformed_headers(self, token):
-        with self.app.test_client() as test_client:
-            response = test_client.get("/api/v1/pools", headers={"Authorization": token})
-            assert response.status_code == 401
-            assert response.headers["Content-Type"] == "application/problem+json"
-            assert response.headers["WWW-Authenticate"] == "Basic"
-            assert_401(response)
-
-    @pytest.mark.parametrize(
-        "token",
-        [
-            "basic " + b64encode(b"test").decode(),
-            "basic " + b64encode(b"test:").decode(),
-            "basic " + b64encode(b"test:123").decode(),
-            "basic " + b64encode(b"test test").decode(),
-        ],
-    )
-    def test_invalid_auth_header(self, token):
-        with self.app.test_client() as test_client:
-            response = test_client.get("/api/v1/pools", headers={"Authorization": token})
-            assert response.status_code == 401
-            assert response.headers["Content-Type"] == "application/problem+json"
-            assert response.headers["WWW-Authenticate"] == "Basic"
-            assert_401(response)
 
 
 class TestSessionAuth(BaseTestAuth):
@@ -144,29 +49,33 @@ class TestSessionAuth(BaseTestAuth):
         finally:
             setattr(minimal_app_for_api, "api_auth", old_auth)
 
-    def test_success(self):
+    @patch.object(SimpleAuthManager, "is_logged_in", return_value=True)
+    @patch.object(
+        SimpleAuthManager, "get_user", return_value=SimpleAuthManagerUser(username="test", role="admin")
+    )
+    def test_success(self, *args):
         clear_db_pools()
 
-        admin_user = client_with_login(self.app, username="test", password="test")
-        response = admin_user.get("/api/v1/pools")
-        assert response.status_code == 200
-        assert response.json == {
-            "pools": [
-                {
-                    "name": "default_pool",
-                    "slots": 128,
-                    "occupied_slots": 0,
-                    "running_slots": 0,
-                    "queued_slots": 0,
-                    "scheduled_slots": 0,
-                    "deferred_slots": 0,
-                    "open_slots": 128,
-                    "description": "Default pool",
-                    "include_deferred": False,
-                },
-            ],
-            "total_entries": 1,
-        }
+        with self.app.test_client() as test_client:
+            response = test_client.get("/api/v1/pools")
+            assert response.status_code == 200
+            assert response.json == {
+                "pools": [
+                    {
+                        "name": "default_pool",
+                        "slots": 128,
+                        "occupied_slots": 0,
+                        "running_slots": 0,
+                        "queued_slots": 0,
+                        "scheduled_slots": 0,
+                        "deferred_slots": 0,
+                        "open_slots": 128,
+                        "description": "Default pool",
+                        "include_deferred": False,
+                    },
+                ],
+                "total_entries": 1,
+            }
 
     def test_failure(self):
         with self.app.test_client() as test_client:
@@ -174,44 +83,3 @@ class TestSessionAuth(BaseTestAuth):
             assert response.status_code == 401
             assert response.headers["Content-Type"] == "application/problem+json"
             assert_401(response)
-
-
-class TestSessionWithBasicAuthFallback(BaseTestAuth):
-    @pytest.fixture(autouse=True, scope="class")
-    def with_basic_auth_backend(self, minimal_app_for_api):
-        from airflow.www.extensions.init_security import init_api_auth
-
-        old_auth = getattr(minimal_app_for_api, "api_auth")
-
-        try:
-            with conf_vars(
-                {
-                    (
-                        "api",
-                        "auth_backends",
-                    ): "airflow.api.auth.backend.session,airflow.providers.fab.auth_manager.api.auth.backend.basic_auth"
-                }
-            ):
-                init_api_auth(minimal_app_for_api)
-                yield
-        finally:
-            setattr(minimal_app_for_api, "api_auth", old_auth)
-
-    def test_basic_auth_fallback(self):
-        token = "Basic " + b64encode(b"test:test").decode()
-        clear_db_pools()
-
-        # request uses session
-        admin_user = client_with_login(self.app, username="test", password="test")
-        response = admin_user.get("/api/v1/pools")
-        assert response.status_code == 200
-
-        # request uses basic auth
-        with self.app.test_client() as test_client:
-            response = test_client.get("/api/v1/pools", headers={"Authorization": token})
-            assert response.status_code == 200
-
-        # request without session or basic auth header
-        with self.app.test_client() as test_client:
-            response = test_client.get("/api/v1/pools")
-            assert response.status_code == 401
